requirements.txt
chardet==5.0.0
python-dotenv==0.21.0
